makes properly

+++ CS3335/hw8/inc_files/arrprint.h
@@ -1,1 +1,1 @@
-void arrprint(int[8]);
+void arrprint(double[8]);
